Updated startServer script

+++ startServer.sh
@@ -3,5 +3,7 @@
 npm install
 
 PID=`ps | grep "node bin/www" | head -n 1 | awk '{print $1}'`
-kill -2 $PID
+if [[ ! -z "$PID"]]; then
+  kill -2 $PID
+fi
 nohup node bin/www & > nohup.out
